AzureFunctions/PredictionsApi/requirements.txt
azure-functions
pypyodbc
pandas
numpy
scikit-learn